* Move types into `mail.types.*` instead of being *_types.py inside client. * Rename AccountType/DomainType to just Account/Domain.
