+++ src/thunderbird_accounts/mail/types/jmap.py
@@ -1,13 +1,15 @@
-from pydantic.alias_generators import to_snake, to_camel
-from pydantic import BaseModel, Field, ConfigDict, model_serializer, model_validator
 import enum
-from typing import NamedTuple, TypedDict, NotRequired, Optional
+from typing import Optional, TypedDict
+
+from pydantic import BaseModel, ConfigDict, Field, model_serializer, model_validator
+from pydantic.alias_generators import to_camel
 from sqlparse.tokens import Literal
 
 
 class BaseSchema(BaseModel):
     """A base schema that allows us to ingest camelCase names to snake_case variables.
     This will serialize it back to camelCase for requests as well."""
+
     model_config = ConfigDict(
         alias_generator=to_camel,
         serialize_by_alias=True,
